Replaced long and ulong with c_long and c_ulong.

+++ Samples/Chap07/Checker3/Checker3.d
@@ -7,6 +7,7 @@ module Checker3;
 
 import core.runtime;
 import core.thread;
+import core.stdc.config;
 import std.conv;
 import std.math;
 import std.range;
@@ -71,7 +72,7 @@ int myWinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine, int
     }
 
     wndclass.lpfnWndProc   = &ChildWndProc;
-    wndclass.cbWndExtra    = int.sizeof;
+    wndclass.cbWndExtra    = c_long.sizeof;
     wndclass.hIcon         = NULL;
     wndclass.lpszClassName = childClass.toUTF16z;
